# backend/openapi.yaml
openapi: 3.0.0
info:
  title: QQM API
  version: 1.0.0
  description: Client / project / infrastructure management API. Generated from
    the actual implemented routes, controllers, and zod validators
    (backend/src/openapi) — not from any prior design document.
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: /
    description: Same-origin backend
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtained from POST /api/auth/login, sent as `Authorization:
        Bearer <token>`."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - message
      required:
        - error
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - per_page
        - total
        - total_pages
...
    EnvironmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Environment"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Environment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
        vpn_resource_id:
          type: string
          nullable: true
          format: uuid
      required:
        - id
        - project_id
        - name
        - description
        - created_at
        - updated_at
        - deleted_at
        - vpn_resource_id
    EnvironmentDetail:
      allOf:
        - $ref: "#/components/schemas/Environment"
        - type: object
          properties:
            project:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
              required:
                - id
                - name
          required:
            - project
...
    AccessMethod:
      type: string
      nullable: true
      enum:
        - ssh
        - rdp
        - telnet
        - web
        - other
        - null
...
    CredentialReference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        server_id:
          type: string
          format: uuid
        label:
          type: string
        reference_location:
          type: string
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        applies_to_access_method:
          $ref: "#/components/schemas/AccessMethod"
      required:
        - id
        - server_id
        - label
        - reference_location
        - notes
        - created_at
        - applies_to_access_method
...
paths:
...
  /api/environments:
    get:
      tags:
        - Environments
      operationId: listEnvironments
      summary: List environments
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - name
              - created_at
              - updated_at
            description: "Default: name"
          required: false
          description: "Default: name"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
            format: uuid
          required: false
          name: project_id
          in: query
      responses:
        "200":
          description: Paginated environments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Environments
      operationId: createEnvironment
      summary: Create an environment (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: string
                  format: uuid
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
              required:
                - project_id
                - name
      responses:
        "201":
          description: Environment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/environments/{id}:
    get:
      tags:
        - Environments
      operationId: getEnvironmentById
      summary: Get an environment by id, with its parent project inlined
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Environments
      operationId: updateEnvironment
      summary: Update an environment (admin only; optimistic lock; project_id is not
        re-parentable)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                vpn_resource_id:
                  type: string
                  nullable: true
                  format: uuid
                updated_at:
                  type: string
                  minLength: 1
              required:
                - updated_at
      responses:
        "200":
          description: Environment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Environments
      operationId: deleteEnvironment
      summary: Soft-delete an environment (admin only; cascades to its servers and
        hard-deletes their credential_references)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Environment soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/environments/{id}/restore:
    post:
      tags:
        - Environments
      operationId: restoreEnvironment
      summary: Restore a soft-deleted environment (admin only; does not
        cascade-restore servers)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Environment restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/credential-references/{id}:
    get:
      tags:
        - CredentialReferences
      operationId: getCredentialReferenceById
      summary: Get a credential reference by id
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Credential reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialReference"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - CredentialReferences
      operationId: updateCredentialReference
      summary: Update a credential reference (admin or member)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  minLength: 1
                reference_location:
                  type: string
                  minLength: 1
                applies_to_access_method:
                  type: string
                  enum:
                    - ssh
                    - rdp
                    - telnet
                    - web
                    - other
                notes:
                  type: string
      responses:
        "200":
          description: Credential reference updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialReference"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - CredentialReferences
      operationId: deleteCredentialReference
      summary: Hard-delete a credential reference (admin only; no soft delete /
        restore for this resource)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Credential reference deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialReference"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...